+++ src/collective/recipe/backup/tests/test_docs.py
@@ -11,7 +11,9 @@ from zope.testing import renormalizing
 # Importing modules so that we can install their eggs in the test buildout.
 import collective.recipe.backup
 import doctest
+import os
 import re
+import tempfile
 import unittest
 import zc.buildout.testing
 import zc.buildout.tests
@@ -33,6 +35,19 @@ checker = renormalizing.RENormalizing([
 ])
 
 
+_dummy, REPOZO_OUTPUT = tempfile.mkstemp()
+REPOZO_SCRIPT_TEXT = """#!/bin/sh
+echo $* >> {0}""".format(REPOZO_OUTPUT)
+
+
+def check_repozo_output():
+    # Print output and empty the file.
+    with open(REPOZO_OUTPUT) as myfile:
+        print(myfile.read())
+    with open(REPOZO_OUTPUT, 'w') as myfile:
+        myfile.write('')
+
+
 def setUp(test):
     zc.buildout.testing.buildoutSetUp(test)
 
@@ -41,6 +56,20 @@ def setUp(test):
 
     # Install any other recipes that should be available in the tests
     zc.buildout.testing.install_develop('zc.recipe.egg', test)
+
+    # Add mock ``bin/repozo`` script:
+    test.globs['write']('bin', 'repozo', REPOZO_SCRIPT_TEXT)
+    test.globs['system']('chmod u+x bin/repozo')
+
+    # Create var directory:
+    test.globs['mkdir']('var')
+
+    # Add some items to the global definitions of the test,
+    # so we can access them from the doc tests.
+    test.globs.update({
+        'check_repozo_output': check_repozo_output,
+        'REPOZO_SCRIPT_TEXT': REPOZO_SCRIPT_TEXT,
+    })
 
 
 def test_suite():